+++ lib/lockbox/audit.rb
@@ -1,8 +1,0 @@
-module Lockbox
-  class Audit < ActiveRecord::Base
-    self.table_name = "lockbox_audits"
-
-    belongs_to :subject, polymorphic: true
-    belongs_to :viewer, polymorphic: true
-  end
-end