+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 - Added support for `master_key` in `previous_versions`
 - Added `Lockbox.rotate` method
 - Improved performance of `migrate` method
-- Added experimental `Lockbox::Audit` model
+- Added generator for audits
 
 ## 0.3.1 (2019-12-26)